change handling of TEXT_DICT to be able to use the same structure as STS
In french, almost every noun comes with a determinant that varies depending on the noun.
By moving the determinant to the dict, we can use the equivalent of "add to the cards exhausted this combat" instead of the litteral translation of exhaust pile which is never used in base sts.
The new "EXHAUST" key corresponds to usage in base sts.